Finish the OAuth flow skeleton for iOS
The skeleton provides minimal requirement to support the authentication flow by prompting in-app
browser, run the OAuth sign in using preset link and return back to the app with access token. There
is no support on PKCE yet as there is no support on server.

The module provide `signin` method to instantiate the flow by prompting in-app browser and return
the access token as the result.