Implement PKCE in getting token in Sign In
In PKCE flow, the native app generate state and verifier. Used for server to
confirm the app is the one requiring for authorization.

The part generating state and verifier lies on native code as it
requires cryptography related library. `newTransaction` in Javascript
returns the corresponding parameter from native generation.

The flow changes as it will now send code challenge and code challenge
method in Sign In flow. When sign in successes and redirect back to
the app, check the return state, and include the verifier in
requesting tokens.

Resolves: #1